requirements.txt
numpy==2.2.6
opencv-python-headless==4.12.0.88
streamlit==1.51.0
pillow==12.0.0
rawpy
